I never thought this day would come! It appears that PMML now compiles and runs as both a 32-bit, and a 64-bit binary! The tests pass, and all examples compile and sound good.
* Adapted union/struct alignments to work in a 64-bit environment
* changed functions in input.c and smfout.c so that fpval could be
defined as either a macro, or as a normal value depending on bit-depth
* Used an int for variable in the pmml_sprintf function in builtin.c

+++ comp/builtin.c
@@ -1249,7 +1249,7 @@ Object  *argv;
     char fchar;
     int  width, base;
     int	 prec;			/* precision: negative if not specified */
-    unsigned long  num;
+    unsigned int  num;
     PmmlFloat  fpnum;
     char buf[MAXPREC+41], *b;	/* buffer for integer-part digits 
 				   or formatted floating-point number.
@@ -1323,8 +1323,8 @@ Object  *argv;
 	case 'd':
 	    base = 10;
 	    get_intarg(num);
-	    if( (long)num < 0 ) {
-		num = - (long)num;
+	    if( (int)num < 0 ) {
+		num = - (int)num;
 		sign = '-';
 	    } 
 	    goto number;
